refactor(agent): split worker domain into run, policy, telemetry sub-domains
