+++ packages/agent/src/worker/agent-resolution.ts
@@ -6,7 +6,7 @@ import {
   type RunInput,
 } from "@openomni/llm";
 import { BuiltinAgentRegistry, type AgentDefinition } from "../agent";
-import type { ToolExecutor, OrchestratorRunInput } from "./run-worker";
+import type { ToolExecutor, OrchestratorRunInput } from "./run/run-worker";
 
 const DEFAULT_MODEL_CONFIG = {
   providerID: "anthropic",
