+++ packages/agent/src/worker/index.ts
@@ -1,48 +1,10 @@
 // Worker domain — agent runtime, execution primitive, policy enforcement, telemetry
 
-export { type ConcurrencyConfig, ConcurrencyGate } from "./concurrency";
-
-export {
-  type PermissionLevel,
-  type PermissionDecision,
-  type PermissionContext,
-  PermissionGate,
-} from "./permission";
-
-export {
-  type RunBudget,
-  type RunState,
-  type BudgetStatus,
-  RunSupervisor,
-} from "./run-supervisor";
-
-export type {
-  OrchestratorConfig,
-  OrchestrationResult,
-  OrchestrationState,
-  ToolExecutor,
-  SessionMode,
-  OrchestratorRunInput,
-} from "./run-worker";
-
-export { RunWorker } from "./run-worker";
+export * from "./run";
+export * from "./policy";
+export * from "./telemetry";
 
 export { type DLQEntry, DeadLetterQueue } from "./dlq";
-
-export {
-  type SummaryTemplate,
-  type SummaryData,
-  SummaryDelivery,
-} from "./summary";
-
-export {
-  type EventMetadata,
-  type QueueMetrics,
-  type RunMetrics,
-  Observability,
-} from "./observability";
-
-export { type AuditEntry, AuditLog } from "./audit";
 
 export {
   resolveAgentDefinition,
@@ -51,6 +13,3 @@ export {
   resolveAgentForWorker,
   fallbackToolExecutor,
 } from "./agent-resolution";
-
-// Re-export from run-worker-sink (no public types, internal utility)
-export {} from "./run-worker-sink";
