+++ packages/agent/src/worker/policy/permission.ts
