+++ packages/agent/src/worker/run/index.ts
@@ -1,0 +1,18 @@
+export type {
+  OrchestratorConfig,
+  OrchestrationResult,
+  OrchestrationState,
+  ToolExecutor,
+  SessionMode,
+  OrchestratorRunInput,
+} from "./run-worker";
+
+export { RunWorker } from "./run-worker";
+
+export {
+  type SummaryTemplate,
+  type SummaryData,
+  SummaryDelivery,
+} from "./summary";
+
+// run-worker-sink has no public exports
