+++ packages/agent/src/worker/telemetry/audit.ts
