+++ packages/agent/src/worker/telemetry/index.ts
@@ -1,0 +1,8 @@
+export {
+  type EventMetadata,
+  type QueueMetrics,
+  type RunMetrics,
+  Observability,
+} from "./observability";
+
+export { type AuditEntry, AuditLog } from "./audit";
