+++ packages/agent/test/e2e/integration.test.ts
@@ -7,7 +7,7 @@ import { type Task } from "../../src/task/types";
 import { Scheduler } from "../../src/trigger/scheduler";
 import { FilesystemWatcher } from "../../src/trigger/watcher";
 import { Router, Dispatcher, Envelope } from "../../src/dispatch";
-import { RunWorker } from "../../src/worker/run-worker";
+import { RunWorker } from "../../src/worker/run/run-worker";
 
 const sleep = (ms: number) =>
   new Promise<void>((resolve) => setTimeout(resolve, ms));
