+++ packages/agent/test/loop/run-worker.test.ts
@@ -3,16 +3,16 @@ import {
   RunWorker,
   OrchestratorConfig,
   OrchestratorRunInput,
-} from "../../src/worker/run-worker";
+} from "../../src/worker/run/run-worker";
 import { TaskManager } from "../../src/task/manager";
 import { Task } from "../../src/task/types";
 import { TaskStorage } from "../../src/task/storage";
 import { Session } from "@openomni/session";
 import type { Sink } from "@openomni/protocol";
-import { Observability } from "../../src/worker/observability";
-import { AuditLog } from "../../src/worker/audit";
+import { Observability } from "../../src/worker/telemetry/observability";
+import { AuditLog } from "../../src/worker/telemetry/audit";
 import { DeadLetterQueue } from "../../src/worker/dlq";
-import { SummaryDelivery } from "../../src/worker/summary";
+import { SummaryDelivery } from "../../src/worker/run/summary";
 
 describe("RunWorker", () => {
   beforeEach(() => {
